DarkEyeAPIv1.0/requirements.txt
aiodns==2.0.0
anyio==3.5.0
asgiref==3.5.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
checkdmarc==4.5.2
click==8.1.2
click-plugins==1.1.1
colorama==0.4.4
cpe==1.2.1
dnspython<=2.0.0
dnstwist==20220131
fastapi==0.75.1
filelock==3.6.0
future==0.18.2
gitdb==4.0.9
gitdb2==4.0.2
GitPython==3.0.6
h11==0.13.0
httptools==0.4.0
idna==2.10
ipwhois==1.2.0
motor==2.5.1
numpy==1.20.1
netaddr==0.8.0
ocsp-checker==1.9.9
pycares==4.1.2
pycparser==2.21
pydig==0.4.0
pydantic==1.9.0
pydnsbl==1.1.4
pymongo==3.12.3
python-dotenv==0.20.0
python-whois==0.7.3
pywhois==0.1
PyYAML==6.0
pillow==9.2.0
pyopenssl==22.0.0
requests==2.27.1
requests-file==1.5.1
shodan==1.27.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
sslyze==5.1.1
starlette==0.17.1
tld==0.12.6
tldextract==3.2.0
truffleHog==2.2.1
truffleHogRegexes==0.0.7
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
watchgod==0.8.2
websockets==10.2
XlsxWriter==3.0.3
